requirements.txt
deepspeed==0.5.9
sentencepiece==0.1.97
sacremoses==0.0.53
datasets==2.12.0
h5py==3.8.0
biopython==1.80
biotite==0.35.0
brotlipy==0.7.0
numpy==1.24.2
obonet==0.3.1
pandas==1.5.3
pickle5==0.0.11
Pillow==9.3.0
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
subword-nmt==0.3.8
tensorboard==2.11.2
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tfrecord==1.14.1
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
huggingface-hub==0.12.0
matplotlib==3.6.3
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xgboost==1.7.3
zipp==3.12.0
lightgbm==3.3.5
xgboost==1.7.3
BeautifulSoup4==4.11.1
requests==2.24.0
gemmi==0.5.8
networkx==3.0
statsmodels==0.14.0
pynvml==11.5.0
charset-normalizer==3.1.0
fair-esm[esmfold]
# OpenFold and its remaining dependency
dllogger @ git+https://github.com/NVIDIA/dllogger.git
bcbio-gff==0.7.0
fsspec==2023.5.0
triton==1.0.0

